Add javascripthon

+++ index.js
@@ -4,6 +4,8 @@ const { sep: slash } = require('path');
 const path = require('path');
 const loaderUtils = require('loader-utils');
 
+const spawn = require('child_process').spawn;
+
 const properName = name => name.replace(/^./, c => c.toUpperCase());
 const listify = array => array.join(', ')
     // make a comma-separated list ending with a '&' separator
@@ -16,13 +18,22 @@ module.exports = function (source) {
             switches: '-b -n -m',
             folder: `.${slash}__javascript__`,
             install: 'pip install transcrypt',
-            python_version: '3.x'
+            python_version: '3.x',
+            sourcemaps: true
         },
         jiphy: {
             switches: '',
             folder: `.${slash}`,
             install: 'pip install jiphy',
-            python_version: '2.x'
+            python_version: '2.x',
+            sourcemaps: false
+        },
+        pj: {
+            switches: '--inline-map -s -',
+            folder: `.${slash}`,
+            install: 'pip install javascripthon',
+            python_version: '3.x',
+            streaming: true
         }
     };
 
@@ -40,31 +51,61 @@ module.exports = function (source) {
 
     const callback = this.async();
 
-    const entry = this._module.resource;
-    //console.log(`py-loader: compiling ${entry} with ${compilerName}...`);
+    if (compiler.streaming) {
+        var child = spawn(compiler.name, compiler.switches.split(' '));
+        child.stdin.write(source);
 
-    const name = path.basename(entry, ".py");
-    const srcDir = path.dirname(entry, ".py");
-
-    if (!entry.toLowerCase().endsWith(".py")) {
-        console.warn("This loader only handles .py files. This could be a problem with your webpack.config.js file. Please add a rule for .py files in your modules entry.");
-        callback(null, source);
-    }
-
-    cmd.get(`${compiler.name} ${compiler.switches} ${srcDir}${slash}${name}.py`, function(err, data, stderr) {
-        if (!err) {
-            const filename = `${srcDir}${slash}${compiler.folder}${slash}${name}.js`;
-            js = fs.readFileSync(filename, "utf8");
-
-            const sourceMapFile = `${srcDir}${slash}${compiler.folder}${slash}extra${slash}sourcemap${slash}${name}.js`;
-            sourceMap = fs.readFileSync(sourceMapFile + ".map", "utf8")
-            fs.unlinkSync(filename);
-            callback(null, js, sourceMap);
-        }
-        else {
+        var data = '';
+        var error = '';
+        child.stdout.on('data', function (js) {
+            data = data + js;
+        });
+        child.stderr.on('data', function (msg) {
+            error = error + msg;
+        });
+        child.on('exit', function () {
+            callback(error, data);
+        });
+        child.on('error', function(err) {
             console.error(`Some error occurred on ${properName(compiler.name)} compiler execution. Have you installed ${properName(compiler.name)}? If not, please run \`${compiler.install}\` (requires Python ${compiler.python_version})`);
             callback(err);
-        }
+        });
+        child.stdin.end();
+    }
+    else {
+        cmd.get(`${compiler.name} ${compiler.switches} ${srcDir}${slash}${name}.py`, function(err, data, stderr) {
 
-    });
+            const entry = this._module.resource;
+            //console.log(`py-loader: compiling ${entry} with ${compilerName}...`);
+
+            const name = path.basename(entry, ".py");
+            const srcDir = path.dirname(entry, ".py");
+
+            if (!entry.toLowerCase().endsWith(".py")) {
+                console.warn("This loader only handles .py files. This could be a problem with your webpack.config.js file. Please add a rule for .py files in your modules entry.");
+                callback(null, source);
+            }
+
+            if (!err) {
+                const filename = `${srcDir}${slash}${compiler.folder}${slash}${name}.js`;
+                js = fs.readFileSync(filename, "utf8");
+                fs.unlinkSync(filename);
+
+                if (compiler.sourcemaps) {
+                    const sourceMapFile = `${srcDir}${slash}${compiler.folder}${slash}extra${slash}sourcemap${slash}${name}.js`;
+                    sourceMap = fs.readFileSync(sourceMapFile + ".map", "utf8")
+                    fs.unlinkSync(sourceMapFile + ".map");
+                    callback(null, js, sourceMap); }
+                else {
+                    callback(null, js);
+                }
+
+            }
+            else {
+                console.error(`Some error occurred on ${properName(compiler.name)} compiler execution. Have you installed ${properName(compiler.name)}? If not, please run \`${compiler.install}\` (requires Python ${compiler.python_version})`);
+                callback(err);
+            }
+
+        });
+    }
 }
